Fixup PositionedOverlay so it doesn't export more than its index
